refactor single arm task.
